completed JDBC campground, altered campground.java, altered park

+++ src/main/java/com/techelevator/dao/CampgroundDAO.java
@@ -5,7 +5,7 @@ import com.techelevator.model.Campground;
 
 public interface CampgroundDAO {
 	
-	List<Campground> getCampgroundsByParkID(int parkID);
+	List<Campground> getCampgroundsByParkID(long parkID);
 	
 }
